Update dummy rpm version to 999999
The old dummy RPM version is 100, which caused issues
when testing upgrade (since 100 is not greater than the
current latest version 302). This worked well, when we had
"0." prefixed with Presto version.

Also removed the assertion in the dummy RPM, because
the newly built dummy RPM simply doesnt have this file.